# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Real-time energy monitoring API providing grid consumption, electricity price, and solar production data.",
        "title": "Sotehus Backend API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/jorgen-simonsson/sotehus-backend"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/data": {
            "get": {
                "description": "Returns current status of grid consumption, electricity price, and solar production",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "energy"
                ],
                "summary": "Get current energy data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/energy/consumed": {
            "get": {
                "description": "Returns the energy consumed in kWh between two timestamps",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "energy"
                ],
                "summary": "Get consumed energy for a period",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start timestamp (RFC3339 format, e.g. 2026-02-01T00:00:00+01:00)",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Stop timestamp (RFC3339 format, e.g. 2026-02-21T00:00:00+01:00)",
                        "name": "stop",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.EnergyConsumedResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid start/stop parameter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to calculate energy consumption",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "InfluxDB not configured",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/energy/cost": {
            "get": {
                "description": "Calculates the actual energy cost for a period, accounting for changing spot prices.\nGroups records into blocks of constant spot price, computes per-block kWh and cost,\nadds configured price components (transfer, energy tax, dynamic and static additions),\nand applies VAT.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "energy"
                ],
                "summary": "Get energy cost for a period",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start timestamp (RFC3339 format, e.g. 2026-02-01T00:00:00+01:00)",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Stop timestamp (RFC3339 format, e.g. 2026-02-21T00:00:00+01:00)",
                        "name": "stop",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.EnergyCostResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid start/stop parameter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to calculate energy cost",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "InfluxDB not configured",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/energy/sold": {
            "get": {
                "description": "Returns the energy sold in kWh between two timestamps",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "energy"
                ],
                "summary": "Get sold energy for a period",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start timestamp (RFC3339 format, e.g. 2026-02-01T00:00:00+01:00)",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Stop timestamp (RFC3339 format, e.g. 2026-02-21T00:00:00+01:00)",
                        "name": "stop",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.EnergySoldResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid start/stop parameter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to calculate energy sold",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "InfluxDB not configured",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/params": {
            "get": {
                "description": "Returns all persistent configuration parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parameters"
                ],
                "summary": "Get all persistent parameters",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/params.PersistentParam"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get parameters",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new persistent parameter. Returns 409 if the key already exists.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parameters"
                ],
                "summary": "Create a new parameter",
                "parameters": [
                    {
                        "description": "Parameter to create",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/params.CreateParamRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/params.PersistentParam"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Parameter with this key already exists",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to create parameter",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/params/{key}": {
            "get": {
                "description": "Returns a single persistent parameter by its key",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parameters"
                ],
                "summary": "Get a parameter by key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Parameter key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/params.PersistentParam"
                        }
                    },
                    "404": {
                        "description": "Parameter not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to get parameter",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates the description and content of a parameter identified by key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parameters"
                ],
                "summary": "Update an existing parameter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Parameter key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated parameter data",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/params.UpdateParamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/params.PersistentParam"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Parameter not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to update parameter",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/solis": {
            "get": {
                "description": "Returns the most recent record from the \"solis\" InfluxDB bucket. household_load is\ncalculated from the power balance: solar_power - battery_power - grid_power, where\ngrid_power is positive when exporting and battery_power is positive when charging.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "energy"
                ],
                "summary": "Get latest Solis inverter data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SolisResponse"
                        }
                    },
                    "404": {
                        "description": "No Solis data found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to query Solis data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "InfluxDB not configured",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/solis/soc": {
            "get": {
                "description": "Returns the battery state of charge averaged into windows of the given size (in minutes)\nacross the requested time range, e.g. a 24h range with am=15 returns 96 data points.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "energy"
                ],
                "summary": "Get battery SOC data points for a period",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start timestamp (RFC3339 format, e.g. 2026-02-01T00:00:00+01:00)",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Stop timestamp (RFC3339 format, e.g. 2026-02-21T00:00:00+01:00)",
                        "name": "stop",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Aggregation window in minutes (e.g. 15)",
                        "name": "am",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.SOCDataPoint"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid start/stop/am parameter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to query SOC data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "InfluxDB not configured",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/timeseries": {
            "get": {
                "description": "Returns statistics about the historical data stored in InfluxDB",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "energy"
                ],
                "summary": "Get time series statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.TimeSeriesResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to query time series data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "InfluxDB not configured",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/version": {
            "get": {
                "description": "Returns the current version of the API",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Get API version",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.VersionResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns the health status of the API",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.HealthResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.EnergyConsumedResponse": {
            "type": "object",
            "properties": {
                "actual_start": {
                    "type": "string",
                    "example": "2026-02-01T00:01:23+01:00"
                },
                "actual_stop": {
                    "type": "string",
                    "example": "2026-02-20T23:59:45+01:00"
                },
                "consumed": {
                    "type": "number",
                    "example": 523.45
                },
                "start": {
                    "type": "string",
                    "example": "2026-02-01T00:00:00+01:00"
                },
                "stop": {
                    "type": "string",
                    "example": "2026-02-21T00:00:00+01:00"
                },
                "unit": {
                    "type": "string",
                    "example": "kWh"
                }
            }
        },
        "api.EnergyCostBlock": {
            "type": "object",
            "properties": {
                "added_prices": {
                    "type": "number",
                    "example": 0.7
                },
                "consumed_kwh": {
                    "type": "number",
                    "example": 12.34
                },
                "cost": {
                    "type": "number",
                    "example": 14.22
                },
                "produced_kwh": {
                    "type": "number",
                    "example": 1.5
                },
                "spot_price": {
                    "type": "number",
                    "example": 0.45
                },
                "start": {
                    "type": "string",
                    "example": "2026-02-22T00:00:00+01:00"
                },
                "stop": {
                    "type": "string",
                    "example": "2026-02-22T01:00:00+01:00"
                },
                "total_price": {
                    "type": "number",
                    "example": 1.15
                }
            }
        },
        "api.EnergyCostResponse": {
            "type": "object",
            "properties": {
                "blocks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.EnergyCostBlock"
                    }
                },
                "cost_before_vat": {
                    "type": "number",
                    "example": 180.73
                },
                "period_start": {
                    "type": "string",
                    "example": "2026-02-22T00:00:00+01:00"
                },
                "period_stop": {
                    "type": "string",
                    "example": "2026-02-22T12:00:00+01:00"
                },
                "production_benefit": {
                    "type": "number",
                    "example": 5.67
                },
                "total_consumed_kwh": {
                    "type": "number",
                    "example": 156.78
                },
                "total_cost": {
                    "type": "number",
                    "example": 220.24
                },
                "total_produced_kwh": {
                    "type": "number",
                    "example": 23.45
                },
                "unit": {
                    "type": "string",
                    "example": "SEK"
                },
                "vat_percent": {
                    "type": "number",
                    "example": 25
                }
            }
        },
        "api.EnergySoldResponse": {
            "type": "object",
            "properties": {
                "actual_start": {
                    "type": "string",
                    "example": "2026-02-01T00:01:23+01:00"
                },
                "actual_stop": {
                    "type": "string",
                    "example": "2026-02-20T23:59:45+01:00"
                },
                "sold": {
                    "type": "number",
                    "example": 123.45
                },
                "start": {
                    "type": "string",
                    "example": "2026-02-01T00:00:00+01:00"
                },
                "stop": {
                    "type": "string",
                    "example": "2026-02-21T00:00:00+01:00"
                },
                "unit": {
                    "type": "string",
                    "example": "kWh"
                }
            }
        },
        "api.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "api.SOCDataPoint": {
            "type": "object",
            "properties": {
                "timestamp": {
                    "type": "string",
                    "example": "2026-02-22T00:15:00+01:00"
                },
                "value": {
                    "type": "number",
                    "example": 78.5
                }
            }
        },
        "api.SolisResponse": {
            "type": "object",
            "properties": {
                "battery_power": {
                    "type": "number",
                    "example": 1234.1
                },
                "battery_soc": {
                    "type": "number",
                    "example": 34
                },
                "grid_power": {
                    "type": "number",
                    "example": 0.23
                },
                "household_load": {
                    "type": "number",
                    "example": 923.56
                },
                "solar_power": {
                    "type": "number",
                    "example": 5450.23
                },
                "timestamp": {
                    "type": "string",
                    "example": "2025-12-07T16:30:00+01:00"
                }
            }
        },
        "api.TimeSeriesResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 523847
                },
                "first": {
                    "type": "string",
                    "example": "2025-11-01T16:10"
                },
                "last": {
                    "type": "string",
                    "example": "2025-12-07T16:30"
                }
            }
        },
        "api.VersionResponse": {
            "type": "object",
            "properties": {
                "version": {
                    "type": "string",
                    "example": "1.2.0"
                }
            }
        },
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "frequency": {
                    "$ref": "#/definitions/models.FrequencyData"
                },
                "grid": {
                    "$ref": "#/definitions/models.GridData"
                },
                "price": {
                    "$ref": "#/definitions/models.PriceData"
                },
                "solar": {
                    "$ref": "#/definitions/models.SolarData"
                }
            }
        },
        "models.FrequencyData": {
            "type": "object",
            "properties": {
                "frequency": {
                    "type": "number",
                    "example": 50.01
                },
                "lastUpdate": {
                    "type": "string",
                    "example": "2025-12-07T16:30:00+01:00"
                },
                "valid": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.GridData": {
            "type": "object",
            "properties": {
                "lastUpdate": {
                    "type": "string",
                    "example": "2025-12-07T16:30:00+01:00"
                },
                "message": {
                    "type": "string",
                    "example": ""
                },
                "power": {
                    "type": "number",
                    "example": 245.5
                },
                "valid": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.PriceData": {
            "type": "object",
            "properties": {
                "lastUpdate": {
                    "type": "string",
                    "example": "2025-12-07T16:15:00+01:00"
                },
                "price": {
                    "type": "number",
                    "example": 1.2
                },
                "valid": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.SolarData": {
            "type": "object",
            "properties": {
                "lastUpdate": {
                    "type": "string",
                    "example": "2025-12-07T15:00:00+01:00"
                },
                "message": {
                    "type": "string",
                    "example": "No sun"
                },
                "power": {
                    "type": "number",
                    "example": 0
                },
                "valid": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "params.CreateParamRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "{\"value\": 0.04}"
                },
                "description": {
                    "type": "string",
                    "example": "Dynamic addition to price"
                },
                "key": {
                    "type": "string",
                    "example": "DynamicAddPrice"
                }
            }
        },
        "params.PersistentParam": {
            "type": "object",
            "properties": {
                "changed": {
                    "type": "string",
                    "example": "2026-01-15T10:30:00+01:00"
                },
                "content": {
                    "type": "string",
                    "example": "{\"value\": 0.04}"
                },
                "description": {
                    "type": "string",
                    "example": "Dynamic addition to price"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "key": {
                    "type": "string",
                    "example": "DynamicAddPrice"
                }
            }
        },
        "params.UpdateParamRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "{\"value\": 0.05}"
                },
                "description": {
                    "type": "string",
                    "example": "Dynamic addition to price"
                }
            }
        }
    }
}